Move `bezierCurve.ts` util functions to lib
