+++ src/lib/bezierCurve.ts
